requirements.txt
# Core Frameworks
torch>=2.5.0
torchvision>=0.20.0
tensorflow>=2.16.0

# Image Processing
opencv-python>=4.10.0
pillow>=10.0.0
scikit-image>=0.22.0
albumentations>=2.0.0

# Segmentation
segment-anything
openslide-bin

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ML Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.14.0

# Utilities
pyyaml>=6.0
h5py>=3.9.0
